net_test

+++ src/camera.c
@@ -5,8 +5,8 @@
 #include "map.h"
 #include "include/raylib.h"
 
-extern const int SCREEN_WIDTH;
-extern const int SCREEN_HEIGHT;
+extern const float SCREEN_WIDTH;
+extern const float SCREEN_HEIGHT;
 
 void move_camera(Camera2D* camera, const Map* map, float pos_x, float pos_y) {
 	if (pos_x < SCREEN_WIDTH / 2.0) {
